Changed Scintilla pointer types to 64 bits - TODO: search proper macro

+++ src/stc/scintilla/include/Scintilla.h
@@ -36,8 +36,8 @@ int Scintilla_LinkLexers();
 typedef ULONG_PTR uptr_t;
 typedef LONG_PTR sptr_t;
 #else
-typedef unsigned long uptr_t;
-typedef long sptr_t;
+typedef unsigned long long uptr_t;
+typedef long long sptr_t;
 #endif
 
 typedef sptr_t (*SciFnDirect)(sptr_t ptr, unsigned int iMessage, uptr_t wParam, sptr_t lParam);
